Refactor the views to be better sorted
