+++ opentech/apply/projects/views/__init__.py
@@ -1,0 +1,2 @@
+from .payment import *  # NOQA
+from .project import *  # NOQA
