+++ opentech/apply/projects/views/project.py
@@ -1,4 +1,3 @@
-import decimal
 from copy import copy
 
 from django.contrib import messages
@@ -6,7 +5,6 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import UserPassesTestMixin
 from django.core.exceptions import PermissionDenied
 from django.db import transaction
-from django.db.models import Q
 from django.http import Http404
 from django.shortcuts import get_object_or_404, redirect
 from django.utils.decorators import method_decorator
@@ -14,7 +12,6 @@ from django.utils.functional import cached_property
 from django.utils.translation import ugettext_lazy as _
 from django.views.generic import (
     CreateView,
-    DeleteView,
     DetailView,
     FormView,
     UpdateView
@@ -27,20 +24,17 @@ from opentech.apply.utils.storage import PrivateMediaView
 from opentech.apply.utils.views import (
     DelegateableView,
     DelegatedViewMixin,
-    ViewDispatcher
+    ViewDispatcher,
 )
 
-from .files import get_files
-from .forms import (
+from ..files import get_files
+from ..forms import (
     ApproveContractForm,
-    ChangePaymentRequestStatusForm,
     CreateApprovalForm,
-    EditPaymentRequestForm,
     ProjectApprovalForm,
     ProjectEditForm,
     RejectionForm,
     RemoveDocumentForm,
-    RequestPaymentForm,
     SelectDocumentForm,
     SetPendingForm,
     StaffUploadContractForm,
@@ -48,53 +42,16 @@ from .forms import (
     UploadContractForm,
     UploadDocumentForm
 )
-from .models import (
-    CHANGES_REQUESTED,
+from ..models import (
     CONTRACTING,
-    DECLINED,
     IN_PROGRESS,
-    PAID,
     PROJECT_STATUS_CHOICES,
-    REQUEST_STATUS_CHOICES,
-    SUBMITTED,
-    UNDER_REVIEW,
     Approval,
     Contract,
     PacketFile,
-    PaymentRequest,
     Project
 )
-
-
-class ContractsMixin:
-    def get_context_data(self, **kwargs):
-        project = self.get_object()
-        contracts = (project.contracts.select_related('approver')
-                                      .order_by('-created_at'))
-
-        latest_contract = self.get_contract_to_approve(contracts)
-
-        contracts = contracts.filter(is_signed=True, approver__isnull=False)
-
-        if latest_contract:
-            contracts = [latest_contract, *contracts]
-
-        context = super().get_context_data(**kwargs)
-        context['latest_contract'] = latest_contract
-        context['contracts'] = contracts
-        return context
-
-    def get_contract_to_approve(self, contracts):
-        """If there's a contract to approve, get that"""
-        latest = contracts.first()
-
-        if not latest:
-            return
-
-        if latest.approver:
-            return
-
-        return latest
+from .payment import RequestPaymentView
 
 
 class SubmissionFilesMixin:
@@ -117,76 +74,26 @@ class SubmissionFilesMixin:
         return context
 
 
-@method_decorator(staff_required, name='dispatch')
-class ApproveContractView(UpdateView):
-    form_class = ApproveContractForm
-    model = Contract
-    pk_url_kwarg = 'contract_pk'
-
-    def dispatch(self, request, *args, **kwargs):
-        self.project = get_object_or_404(Project, pk=self.kwargs['pk'])
-        return super().dispatch(request, *args, **kwargs)
-
-    def form_invalid(self, form):
-        for error in form.errors:
-            messages.error(self.request, error)
-
-        return redirect(self.project)
-
-    def form_valid(self, form):
-        with transaction.atomic():
-            form.instance.approver = self.request.user
-            form.instance.project = self.project
-            response = super().form_valid(form)
-
-            messenger(
-                MESSAGES.APPROVE_CONTRACT,
-                request=self.request,
-                user=self.request.user,
-                source=self.project,
-                related=self.object,
-            )
-
-            self.project.status = IN_PROGRESS
-            self.project.save(update_fields=['status'])
-
-        return response
-
-    def get_success_url(self):
-        return self.project.get_absolute_url()
-
+# APPROVAL VIEWS
 
 @method_decorator(staff_required, name='dispatch')
-class ChangePaymentRequestStatusView(UpdateView):
-    form_class = ChangePaymentRequestStatusForm
-    model = PaymentRequest
-    pk_url_kwarg = 'payment_request_id'
-
-    def dispatch(self, request, *args, **kwargs):
-        self.project = get_object_or_404(Project, pk=self.kwargs['pk'])
-        return super().dispatch(request, *args, **kwargs)
-
-    def form_invalid(self, form):
-        for error in form.errors:
-            messages.error(self.request, error)
-
-        return redirect(self.project)
+class SendForApprovalView(DelegatedViewMixin, UpdateView):
+    context_name = 'request_approval_form'
+    form_class = SetPendingForm
+    model = Project
 
     def form_valid(self, form):
+        # lock project
         response = super().form_valid(form)
 
         messenger(
-            MESSAGES.UPDATE_PAYMENT_REQUEST_STATUS,
+            MESSAGES.SEND_FOR_APPROVAL,
             request=self.request,
             user=self.request.user,
-            source=self.project,
-            related=self.object,
+            source=self.object,
         )
 
         return response
-
-    def get_success_url(self):
-        return self.project.get_absolute_url()
 
 
 @method_decorator(staff_required, name='dispatch')
@@ -224,112 +131,6 @@ class CreateApprovalView(DelegatedViewMixin, CreateView):
         return response
 
 
-class DeletePaymentRequestView(DeleteView):
-    model = PaymentRequest
-    pk_url_kwarg = 'payment_request_id'
-
-    def dispatch(self, request, *args, **kwargs):
-        self.project = get_object_or_404(Project, pk=self.kwargs['pk'])
-
-        self.object = self.get_object()
-        if not self.object.user_can_delete(request.user):
-            raise PermissionDenied
-
-        return super().dispatch(request, *args, **kwargs)
-
-    @transaction.atomic()
-    def delete(self, request, *args, **kwargs):
-        response = super().delete(request, *args, **kwargs)
-
-        messenger(
-            MESSAGES.DELETE_PAYMENT_REQUEST,
-            request=self.request,
-            user=self.request.user,
-            source=self.project,
-            related=self.object,
-        )
-
-        return response
-
-    def get_success_url(self):
-        return self.project.get_absolute_url()
-
-
-@method_decorator(login_required, name='dispatch')
-class PaymentRequestAccessMixin:
-    model = PaymentRequest
-    pk_url_kwarg = 'pr_pk'
-
-    def dispatch(self, request, *args, **kwargs):
-        self.project = get_object_or_404(Project, pk=self.kwargs['pk'])
-        if self.get_object().project != self.project:
-            raise Http404
-
-        is_admin = request.user.is_apply_staff
-        is_owner = request.user == self.project.user
-        if not (is_owner or is_admin):
-            raise PermissionDenied
-
-        return super().dispatch(request, *args, **kwargs)
-
-
-class PaymentRequestView(PaymentRequestAccessMixin, DetailView):
-    pass
-
-
-class EditPaymentRequestView(PaymentRequestAccessMixin, UpdateView):
-    form_class = EditPaymentRequestForm
-
-    def form_valid(self, form):
-        response = super().form_valid(form)
-
-        messenger(
-            MESSAGES.UPDATE_PAYMENT_REQUEST,
-            request=self.request,
-            user=self.request.user,
-            source=self.project,
-            related=self.object,
-        )
-
-        return response
-
-    def get_success_url(self):
-        return self.project.get_absolute_url()
-
-
-@method_decorator(login_required, name='dispatch')
-class PaymentRequestPrivateMedia(UserPassesTestMixin, PrivateMediaView):
-    raise_exception = True
-
-    def dispatch(self, *args, **kwargs):
-        project_pk = self.kwargs['pk']
-        payment_pk = self.kwargs['pr_pk']
-        self.project = get_object_or_404(Project, pk=project_pk)
-        self.payment_request = get_object_or_404(PaymentRequest, pk=payment_pk)
-
-        if self.payment_request.project != self.project:
-            raise Http404
-
-        return super().dispatch(*args, **kwargs)
-
-    def get_media(self, *args, **kwargs):
-        file_pk = kwargs.get('file_pk')
-        if not file_pk:
-            return self.payment_request.invoice
-
-        receipt = get_object_or_404(self.payment_request.receipts, pk=file_pk)
-        return receipt.file
-
-    def test_func(self):
-        if self.request.user.is_apply_staff:
-            return True
-
-        if self.request.user == self.project.user:
-            return True
-
-        return False
-
-
 @method_decorator(staff_required, name='dispatch')
 class RejectionView(DelegatedViewMixin, UpdateView):
     context_name = 'rejection_form'
@@ -351,6 +152,29 @@ class RejectionView(DelegatedViewMixin, UpdateView):
         return redirect(self.object)
 
 
+# PROJECT DOCUMENTS
+
+@method_decorator(staff_required, name='dispatch')
+class UploadDocumentView(DelegatedViewMixin, CreateView):
+    context_name = 'document_form'
+    form_class = UploadDocumentForm
+    model = Project
+
+    def form_valid(self, form):
+        project = self.kwargs['object']
+        form.instance.project = project
+        response = super().form_valid(form)
+
+        messenger(
+            MESSAGES.UPLOAD_DOCUMENT,
+            request=self.request,
+            user=self.request.user,
+            source=project,
+        )
+
+        return response
+
+
 @method_decorator(staff_required, name='dispatch')
 class RemoveDocumentView(DelegatedViewMixin, FormView):
     context_name = 'remove_document_form'
@@ -367,28 +191,6 @@ class RemoveDocumentView(DelegatedViewMixin, FormView):
             pass
 
         return redirect(project)
-
-
-class RequestPaymentView(DelegatedViewMixin, CreateView):
-    context_name = 'request_payment_form'
-    form_class = RequestPaymentForm
-    model = PaymentRequest
-
-    def form_valid(self, form):
-        project = self.kwargs['object']
-
-        form.instance.by = self.request.user
-        form.instance.project = project
-        response = super().form_valid(form)
-
-        messenger(
-            MESSAGES.APPROVE_PROJECT,
-            request=self.request,
-            user=self.request.user,
-            source=project,
-        )
-
-        return response
 
 
 @method_decorator(login_required, name='dispatch')
@@ -430,25 +232,7 @@ class SelectDocumentView(SubmissionFilesMixin, FormView):
         return self.project.get_absolute_url()
 
 
-@method_decorator(staff_required, name='dispatch')
-class SendForApprovalView(DelegatedViewMixin, UpdateView):
-    context_name = 'request_approval_form'
-    form_class = SetPendingForm
-    model = Project
-
-    def form_valid(self, form):
-        # lock project
-        response = super().form_valid(form)
-
-        messenger(
-            MESSAGES.SEND_FOR_APPROVAL,
-            request=self.request,
-            user=self.request.user,
-            source=self.object,
-        )
-
-        return response
-
+# GENERAL FORM VIEWS
 
 @method_decorator(staff_required, name='dispatch')
 class UpdateLeadView(DelegatedViewMixin, UpdateView):
@@ -471,6 +255,78 @@ class UpdateLeadView(DelegatedViewMixin, UpdateView):
         )
 
         return response
+
+
+# CONTRACTS
+
+class ContractsMixin:
+    def get_context_data(self, **kwargs):
+        project = self.get_object()
+        contracts = (project.contracts.select_related('approver')
+                                      .order_by('-created_at'))
+
+        latest_contract = self.get_contract_to_approve(contracts)
+
+        contracts = contracts.filter(is_signed=True, approver__isnull=False)
+
+        if latest_contract:
+            contracts = [latest_contract, *contracts]
+
+        context = super().get_context_data(**kwargs)
+        context['latest_contract'] = latest_contract
+        context['contracts'] = contracts
+        return context
+
+    def get_contract_to_approve(self, contracts):
+        """If there's a contract to approve, get that"""
+        latest = contracts.first()
+
+        if not latest:
+            return
+
+        if latest.approver:
+            return
+
+        return latest
+
+
+@method_decorator(staff_required, name='dispatch')
+class ApproveContractView(UpdateView):
+    form_class = ApproveContractForm
+    model = Contract
+    pk_url_kwarg = 'contract_pk'
+
+    def dispatch(self, request, *args, **kwargs):
+        self.project = get_object_or_404(Project, pk=self.kwargs['pk'])
+        return super().dispatch(request, *args, **kwargs)
+
+    def form_invalid(self, form):
+        for error in form.errors:
+            messages.error(self.request, error)
+
+        return redirect(self.project)
+
+    def form_valid(self, form):
+        with transaction.atomic():
+            form.instance.approver = self.request.user
+            form.instance.project = self.project
+            response = super().form_valid(form)
+
+            messenger(
+                MESSAGES.APPROVE_CONTRACT,
+                request=self.request,
+                user=self.request.user,
+                source=self.project,
+                related=self.object,
+            )
+
+            self.project.status = IN_PROGRESS
+            self.project.save(update_fields=['status'])
+
+        return response
+
+    def get_success_url(self):
+        return self.project.get_absolute_url()
 
 
 @method_decorator(login_required, name='dispatch')
@@ -519,26 +375,7 @@ class UploadContractView(DelegatedViewMixin, CreateView):
         return response
 
 
-@method_decorator(staff_required, name='dispatch')
-class UploadDocumentView(DelegatedViewMixin, CreateView):
-    context_name = 'document_form'
-    form_class = UploadDocumentForm
-    model = Project
-
-    def form_valid(self, form):
-        project = self.kwargs['object']
-        form.instance.project = project
-        response = super().form_valid(form)
-
-        messenger(
-            MESSAGES.UPLOAD_DOCUMENT,
-            request=self.request,
-            user=self.request.user,
-            source=project,
-        )
-
-        return response
-
+# PROJECT VIEW
 
 class AdminProjectDetailView(
     ActivityContextMixin,
@@ -567,20 +404,8 @@ class AdminProjectDetailView(
         context['current_status_index'] = [status for status, _ in PROJECT_STATUS_CHOICES].index(self.object.status)
         context['approvals'] = self.object.approvals.distinct('by')
         context['approve_contract_form'] = ApproveContractForm()
-        context['change_payment_request_status_forms'] = list(self.get_change_payment_request_status_forms())
         context['remaining_document_categories'] = list(self.object.get_missing_document_categories())
         return context
-
-    def get_change_payment_request_status_forms(self):
-        """
-        Get an iterable of ChangePaymentRequestStatusForms
-
-        We want to filter the available options based on the current
-        PaymentRequest object so we need to initialise those forms outside of
-        the template.
-        """
-        for payment_request in self.object.payment_requests.in_progress().select_related('project'):
-            yield ChangePaymentRequestStatusForm(instance=payment_request)
 
 
 class ApplicantProjectDetailView(ActivityContextMixin, DelegateableView, ContractsMixin, DetailView):
@@ -598,6 +423,11 @@ class ApplicantProjectDetailView(ActivityContextMixin, DelegateableView, Contrac
         if project.user != request.user:
             raise PermissionDenied
         return super().dispatch(request, *args, **kwargs)
+
+
+class ProjectDetailView(ViewDispatcher):
+    admin_view = AdminProjectDetailView
+    applicant_view = ApplicantProjectDetailView
 
 
 @method_decorator(login_required, name='dispatch')
@@ -625,10 +455,7 @@ class ProjectPrivateMediaView(UserPassesTestMixin, PrivateMediaView):
         return False
 
 
-class ProjectDetailView(ViewDispatcher):
-    admin_view = AdminProjectDetailView
-    applicant_view = ApplicantProjectDetailView
-
+# PROJECT EDIT
 
 @method_decorator(staff_required, name='dispatch')
 class ProjectDetailSimplifiedView(DetailView):
